+++ src/scl/utility.h
@@ -7,6 +7,7 @@
 
 #include <scl/utility/attribute.h>
 #include <scl/utility/concepts.h>
+#include <scl/utility/hierarchy.h>
 #include <scl/utility/meta.h>
 #include <scl/utility/preprocessor.h>
 #include <scl/utility/type_traits.h>